+++ changelog.md
@@ -3,6 +3,7 @@
 #### v3.5.1
 
 * `[usage]` Using dark grey color for license and copyright
+* `[fmtutil]` Added global variable `SeparatorColorTag` for separator color customization
 
 #### v3.5.0